ServiceManager: Allow system services running
as secondary users to add services

This should be reverted when all system
services have been cleaned up to not
do this. A process looking up a service
while running in the background will see
the service registered by the active user
(assuming the service is
registered on every user switch), not the
 service registered by the user that
the process itself belongs to.

BUG: 30795333
Change-Id: I1b74d58be38ed358f43c163692f9e704f8f31dbe
(cherry picked from commit e6bbe69ba739c8a08837134437aaccfea5f1d943)